ci: more cjs for website
